Split loot modifiers to sperate files
